refactor(mobile,connection): relocate device-connection hooks file + collapse useEffect into module-level QueryCache subscription (OJD-1522)
W2 of the component-quality refactor.

File move:
  features/connection/services/device-connection-manager/device-connection-hooks.ts
  -> features/connection/hooks/use-device-connection.ts

The file calls useQuery, useQueryClient, useState, and exports
useConnectedDevice / useConnectToDevice / useAllDevices /
usePairedDevices / useSerialDevices — it's a hooks file, not a
service. Convention rule 4 places it under hooks/.

Effect removal:
The useConnectedDevice hook used to keep a per-consumer useEffect
(L52-59) that watched its useQuery data, detected non-null -> null
transitions (i.e. disconnects the native onDeviceDisconnected event
missed on Android), and cleared the scanner-executor store. Each
consumer paid that effect.

Replaced with one module-level QueryCache subscription bound by the
existing initConnectedDeviceListeners (renamed from
initDisconnectListener). The subscription watches the
connected-device query key globally and runs the scanner-store
cleanup exactly when data transitions non-null -> null — whether the
trigger was the native event or the 3s polling fallback. One listener
for the whole app instead of one per consumer.

Native onDeviceDisconnected now only invalidates the query; the
QueryCache subscriber handles the cleanup uniformly. Removes the
'native handler clears + invalidates, then refetch flips to null, then
subscriber would also clear' double-clear path by routing through one
code path.

Import paths updated in 5 callers:
  - shared/ui/widgets/device-connection-widget.tsx
  - features/measurement-flow/.../measurement-node/measurement-node.tsx
  - features/measurement-flow/.../measurement-node/components/no-device-state.tsx
  - features/connection/hooks/use-auto-reconnect.ts
  - features/connection/components/connection-setup/connection-setup.tsx

Verification: typecheck 0 errors, lint clean, 383/383 tests pass.
