+++ apps/mobile/src/features/connection/hooks/use-device-connection.ts
@@ -1,64 +1,79 @@
 import { QueryClient, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import RNBluetoothClassic from "react-native-bluetooth-classic";
 import { useDeviceConnectionStore } from "~/features/connection/hooks/use-device-connection-store";
 import { useScannerCommandExecutorStore } from "~/features/connection/stores/use-scanner-command-executor-store";
 import type { Device } from "~/shared/types/device";
 
-import { connectToDevice, disconnectFromDevice, unpairDevice } from "./device-connection";
+import {
+  connectToDevice,
+  disconnectFromDevice,
+  unpairDevice,
+} from "../services/device-connection-manager/device-connection";
 import {
   getConnectedDevice,
   getAllDevices,
   getPairedDevices,
   getSerialDevices,
-} from "./device-queries";
+} from "../services/device-connection-manager/device-queries";
+
+const CONNECTED_DEVICE_KEY = ["connected-device"] as const;
 
 /**
- * Register the native disconnect listener once at module level so that
- * multiple components calling useConnectedDevice() don't each create their
- * own subscription.  The listener is bound lazily on the first call to
- * initDisconnectListener and never removed (lives for the app lifetime).
+ * Bind once-per-app-lifetime listeners that keep the scanner executor in
+ * sync with the connected-device cache:
+ *
+ *   1. Native onDeviceDisconnected — fires immediately when the OS reports
+ *      a disconnect (when it bothers to). Invalidates the query so the UI
+ *      flips to the disconnect state ASAP.
+ *   2. QueryCache subscriber on the connected-device key — catches the
+ *      polling-detected disconnect case (common on Android when the device
+ *      is simply powered off and no native event fires) AND the native
+ *      event case after the invalidation refetches null. Either way, the
+ *      scanner executor store gets cleared exactly when data transitions
+ *      from non-null → null.
+ *
+ * Previously the QueryCache-subscriber path lived as a useEffect inside
+ * useConnectedDevice, which ran once per consumer. The module-level
+ * subscription is one listener for the whole app, and removes the only
+ * useEffect that lived in this hook file.
  */
-let disconnectListenerBound = false;
-function initDisconnectListener(client: QueryClient) {
-  if (disconnectListenerBound) return;
-  disconnectListenerBound = true;
+let listenersBound = false;
+function initConnectedDeviceListeners(client: QueryClient) {
+  if (listenersBound) return;
+  listenersBound = true;
+
   RNBluetoothClassic.onDeviceDisconnected(() => {
-    // Clean up the scanner executor so the stale connection doesn't block
-    // the next connectToDevice → setDevice call (isInitializing guard).
-    void useScannerCommandExecutorStore.getState().setDevice(undefined);
-    void client.invalidateQueries({ queryKey: ["connected-device"] });
+    void client.invalidateQueries({ queryKey: CONNECTED_DEVICE_KEY });
+  });
+
+  let prev: Device | null | undefined = client.getQueryData(CONNECTED_DEVICE_KEY);
+  client.getQueryCache().subscribe((event) => {
+    if (event.type !== "updated") return;
+    if (event.query.queryKey[0] !== CONNECTED_DEVICE_KEY[0]) return;
+    const next = event.query.state.data as Device | null | undefined;
+    if (prev && !next) {
+      void useScannerCommandExecutorStore.getState().setDevice(undefined);
+    }
+    prev = next;
   });
 }
 
 export function useConnectedDevice() {
   const client = useQueryClient();
-  initDisconnectListener(client);
+  initConnectedDeviceListeners(client);
 
-  const { data, isLoading, error } = useQuery({
-    queryKey: ["connected-device"],
+  return useQuery({
+    queryKey: CONNECTED_DEVICE_KEY,
     queryFn: getConnectedDevice,
     networkMode: "always",
     // Poll so we catch disconnects even when the native
     // onDeviceDisconnected event doesn't fire (common on Android
-    // when the device is simply powered off).
+    // when the device is simply powered off). The module-level
+    // QueryCache subscriber turns these polling-detected transitions
+    // into the scanner-store cleanup.
     refetchInterval: 3000,
   });
-
-  // When polling detects a disconnect that the native listener missed,
-  // clean up the command executor so the measurement page shows the
-  // reconnect prompt instead of letting the user attempt a scan.
-  const prevDeviceRef = useRef(data);
-  useEffect(() => {
-    const prev = prevDeviceRef.current;
-    prevDeviceRef.current = data;
-
-    if (prev && !data) {
-      void useScannerCommandExecutorStore.getState().setDevice(undefined);
-    }
-  }, [data]);
-
-  return { data, isLoading, error };
 }
 
 export function useConnectToDevice() {
