+++ apps/mobile/src/features/measurement-flow/screens/measurement-flow-screen/components/flow-nodes/measurement-node/measurement-node.tsx
@@ -4,8 +4,8 @@ import { Info } from "lucide-react-native";
 import React, { useEffect, useRef } from "react";
 import { View, Text } from "react-native";
 import { toast } from "sonner-native";
+import { useConnectedDevice } from "~/features/connection/hooks/use-device-connection";
 import { useScanner } from "~/features/connection/hooks/use-scan-manager";
-import { useConnectedDevice } from "~/features/connection/services/device-connection-manager/device-connection-hooks";
 import { useProtocol } from "~/features/measurement-flow/hooks/use-protocol";
 import { useMeasurementFlowStore } from "~/features/measurement-flow/stores/use-measurement-flow-store";
 import { useTranslation } from "~/shared/i18n";
