+++ apps/mobile/src/shared/ui/widgets/device-connection-widget.tsx
@@ -4,9 +4,9 @@ import { Unplug, RefreshCw, WifiOff } from "lucide-react-native";
 import { useState } from "react";
 import { View, Text, TouchableOpacity, ActivityIndicator } from "react-native";
 import { toast } from "sonner-native";
+import { useConnectedDevice } from "~/features/connection/hooks/use-device-connection";
 import { useDeviceConnectionStore } from "~/features/connection/hooks/use-device-connection-store";
 import { useScannerCommandExecutor } from "~/features/connection/hooks/use-scanner-command-executor";
-import { useConnectedDevice } from "~/features/connection/services/device-connection-manager/device-connection-hooks";
 import { useIsOnline } from "~/shared/ui/hooks/use-is-online";
 import { useTheme } from "~/shared/ui/hooks/use-theme";
 import { isOnline } from "~/shared/utils/is-online";
